% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrtruncnorm.R
\name{altrtruncnorm}
\alias{altrtruncnorm}
\title{Generate trancated random normal variables where values outside range are assigned to range limit}
\usage{
altrtruncnorm(n, a = 0, b = 1, meana = 0, sda = 1)
}
\arguments{
\item{n}{number of variables to be generated}

\item{a}{lower limit of range}

\item{b}{upper limit of range}

\item{meana}{mean of underlying normal distribution prior to truncation}

\item{sda}{standard deviation of underlying normal distribution prior to truncation}
}
\description{
Note that this function is different from the one in the package truncnorm
}
\details{
Updated 2018-08-14
}
\examples{
truncnorm()
}
\keyword{misc}
