% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesim_app.R
\name{onesim_app}
\alias{onesim_app}
\title{Generate an interactive Shiny interface for one simulation}
\usage{
onesim_app(pHSinit = 0.8, Kx = 100, betax = 0.02, wxtnormm = 0.8,
  wxtnormsd = 0.3, hx = 1, mxtnormm = 1, mxtnormsd = 0.1,
  axtnormm = 1, axtnormsd = 0.1, rx = 0.1, zxtnormm = 1,
  zxtnormsd = 0.1, gx = 4, cx = 0.9, phix = 0, nseasons = 10,
  HPcut = 0.5, pHScut = 0.5, maY = 100, miY = 0, thetax = 0.2,
  Ex = 0)
}
\arguments{
\item{pHSinit}{the initial proportion of healthy seed, numeric or numeric vector.}

\item{Kx}{the total number of plants, positive interger, numeric or numeric vector.}

\item{betax}{the maximum seasonal transmission rate, numeric or numeric vector.}

\item{wxtnormm}{the environmental effect on transmission rate (mean of underlying normal distribution prior to truncation), numeric or numeric vector.}

\item{wxtnormsd}{the environmental effect on transmission rate (standard deviation of underlying normal distribution prior to truncation), numeric or numeric vector.}

\item{hx}{the host effect on transmission rate, numeric or numeric vector.}

\item{mxtnormm}{the vector management effect on transmission rate (mean of underlying normal distribution prior to truncation), numeric or numeric vector.}

\item{mxtnormsd}{the vector management effect on transmission rate (standard deviation of underlying normal distribution prior to truncation), numeric or numeric vector.}

\item{axtnormm}{the roguing effect in terms of decreased DP (mean of underlying normal distribution prior to truncation), numeric or numeric vector.}

\item{axtnormsd}{the roguing effect in terms of decreased DP (standard deviation of underlying normal distribution prior to truncation), numeric or numeric vector.}

\item{rx}{the reversion rate, numeric or numeric vector.}

\item{zxtnormm}{the proportional selection against diseased plants (mean of underlying normal distribution prior to truncation), numeric or numeric vector.}

\item{zxtnormsd}{the proportional selection against diseased plants (standard deviation of underlying normal distribution prior to truncation), numeric or numeric vector.}

\item{gx}{the seed production rate in healthy plants, numeric or numeric vector.}

\item{cx}{the proportional seed production rate in diseased plants, numeric or numeric vector.}

\item{phix}{the proportion clean seed purchased, numeric or numeric vector.}

\item{nseasons}{the number of seasons, numeric or numeric vector.}

\item{HPcut}{the proportion healthy plant number cutoff, numeric or numeric vector.}

\item{pHScut}{the proportion healthy seed cutoff, numeric or numeric vector.}

\item{maY}{the maximum attainable yield, end of season, in the absence of disease, numeric or numeric vector.}

\item{miY}{the minimum yield when all plants are diseased (useable yield despite disease), numeric or numeric vector.}

\item{thetax}{the rate of decline of Yld with increasing disease incidence, numeric or numeric vector.}

\item{Ex}{the amount of external inoculum around field, numeric or numeric vector.}
}
\description{
This function simulates one parameter combination across nseasons once
}
\details{
Updated 2018-08-23
This Shiny application interface is generated by function \code{\link{onesim}}.
}
\examples{
onesim_app()
}
\keyword{health}
\keyword{seed}
