% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasModdet.R
\name{SeasModdet}
\alias{SeasModdet}
\title{Generate an output matrix for one simulation}
\usage{
SeasModdet(pHSinit = 0.8, Kx = 100, betax = 0.02, wx = 0.8,
  hx = 1, mx = 1, ax = 1, rx = 0.1, zx = 0.9, gx = 4,
  cx = 0.9, phix = 0, maY = 100, miY = 0, thetax = 0.2,
  nseasons = 10, Ex = 0)
}
\arguments{
\item{Kx}{total number of plants}

\item{betax}{maximum seasonal transmission rate}

\item{wx}{environmental effect on transmission rate}

\item{hx}{host effect on transmission rate}

\item{mx}{vector management effect on transmission rate}

\item{ax}{roguing effect in terms of decreased DP}

\item{rx}{reversion rate}

\item{zx}{proportional selection against diseased plants}

\item{gx}{seed production rate in healthy plants}

\item{cx}{proportional seed production rate in diseased plants}

\item{phix}{proportion clean seed purchased}

\item{maY}{maximum attainable yield, end of season, in the absence of disease}

\item{miY}{minimum yield when all plants are diseased (useable yield despite disease)}

\item{thetax}{rate of decline of Yld with increasing disease incidence}

\item{nseasons}{number of seasons}

\item{Ex}{amount of external inoculum around field}
}
\description{
This function simulates one parameter combination across nseasons once
}
\details{
Updated 2018-08-14
}
\examples{
sim1()
sim1()
}
\keyword{health}
\keyword{seed}
