% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasModstoch.R
\name{SeasModstoch}
\alias{SeasModstoch}
\title{Generate an output matrix for one simulation}
\usage{
SeasModstoch(pHSinit = 0.8, Kx = 100, betax = 0.02, wxtnormm = 0.8,
  wxtnormsd = 0.3, hx = 1, mxtnormm = 1, mxtnormsd = 0.1,
  axtnormm = 1, axtnormsd = 0.1, rx = 0.1, zxtnormm = 1,
  zxtnormsd = 0.1, gx = 4, cx = 0.9, phix = 0, nseasons = 10,
  HPcut = 0.5, pHScut = 0.5, maY = 100, miY = 0, thetax = 0.2,
  Ex = 0)
}
\arguments{
\item{Kx}{total number of plants}

\item{betax}{maximum seasonal transmission rate}

\item{hx}{host effect on transmission rate}

\item{rx}{reversion rate}

\item{gx}{seed production rate in healthy plants}

\item{cx}{proportional seed production rate in diseased plants}

\item{phix}{proportion clean seed purchased}

\item{nseasons}{number of seasons}

\item{maY}{maximum attainable yield, end of season, in the absence of disease}

\item{miY}{minimum yield when all plants are diseased (useable yield despite disease)}

\item{thetax}{rate of decline of Yld with increasing disease incidence}

\item{Ex}{amount of external inoculum around field}

\item{wx}{environmental effect on transmission rate}

\item{mx}{vector management effect on transmission rate}

\item{ax}{roguing effect in terms of decreased DP}

\item{zx}{proportional selection against diseased plants}
}
\description{
This function simulates one parameter combination across nseasons once
}
\details{
Updated 2018-08-14
}
\examples{
sim1()
sim1()
}
\keyword{health}
\keyword{seed}
