% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{mpc}
\alias{mpc}
\title{Create a matrix with output for multiple parameter combinations}
\usage{
mpc(vpHSinit = c(0.2, 0.5, 0.8), vKx = 100, vbetax = c(0.02, 0.04),
  vwxtnormm = c(0.3, 0.7), vwxtnormsd = c(0.3, 0.1), vhx = 1,
  vmxtnormm = 1, vmxtnormsd = 0.1, vaxtnormm = c(1, 0.5),
  vaxtnormsd = c(0.3, 0.1), vrx = c(0.1, 0.3), vzxtnormm = c(1, 0.5),
  vzxtnormsd = c(0.3, 0.1), vgx = 4, vcx = 0.9, vphix = c(0, 0.5),
  vnseasons = 10, vnsim = 1, vHPcut = 0.5, vpHScut = 0.5,
  vmaY = 100, vmiY = 0, vthetax = c(-0.5, 0, 0.5), vEx = 0.02)
}
\arguments{
\item{Kx}{total number of plants}

\item{betax}{maximum seasonal transmission rate}

\item{wx}{environmental effect on transmission rate}

\item{hx}{host effect on transmission rate}

\item{mx}{vector management effect on transmission rate}

\item{ax}{roguing effect in terms of decreased DP}

\item{maY}{maximum attainable yield, end of season, in the absence of disease}

\item{miY}{minimum yield when all plants are diseased (useable yield despite disease)}

\item{thetax}{rate of decline of Yld with increasing disease incidence}

\item{Ex}{amount of external inoculum around field}

\item{rx}{reversion rate}

\item{zx}{proportional selection against diseased plants}

\item{gx}{seed production rate in healthy plants}

\item{cx}{proportional seed production rate in diseased plants}

\item{phix}{proportion clean seed purchased}

\item{nseasons}{number of seasons}
}
\description{
This function creates a matrix with output for multiple parameter combinations across nseasons
}
\details{
Updated 2018-08-21
}
\examples{
Multstoch()
}
\keyword{health}
\keyword{seed}
